requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
pyyaml==6.0.2
python-dotenv==1.0.1
secure-smtplib==0.1.1